edit: changed phrasing and added tests to 2 rules
Based on feedback from Notion about what the page permission
events mean, we adjusted the names/IDs of the related rules. Also
added some unit tests and tweaked the titles/alert_contexts.
