+++ packs/notion.yml
@@ -5,12 +5,12 @@ PackDefinition:
   IDs:
     - Notion.AccountChangedAfterLogin
     - Notion.Audit.Log.Exported
-    - Notion.GuestAddedToPage
+    - Notion.PagePerms.GuestPermsChanged
     - Notion.LoginFromBlockedIP
     - Notion.LoginFromNewLocation
     - Notion.Many.Pages.Exported
     - Notion.Many.Pages.Deleted
-    - Notion.PageAccessibleToAPI
+    - Notion.PagePerms.APIPermsChanged
     - Notion.PageViews.ImpossibleTravel
     - Notion.PageSharedToWeb
     - Notion.SAML.SSO.Configuration.Changed
