--- rules/notion_rules/notion_guest_added_to_page.yml
@@ -1,16 +1,0 @@
-AnalysisType: rule
-RuleID: "Notion.GuestAddedToPage"
-DisplayName: "Notion Guest Added to Page"
-Description: An external guest was added to a Notion page.
-LogTypes:
-    - Notion.AuditLogs
-Enabled: true
-Filename: notion_guest_added_to_page.py
-Runbook: Potential information exposure - review the shared page and rectify if needed.
-Severity: Low
-DedupPeriodMinutes: 60
-Threshold: 1
-Tags:
-  - Notion
-  - Public Exposure
-  - Beta
