+++ rules/notion_rules/notion_page_accessible_to_api.yml
@@ -1,7 +1,7 @@
 AnalysisType: rule
-RuleID: "Notion.PageAccessibleToAPI"
-DisplayName: "Notion Page Shared with API"
-Description: An API integration was added to a Notion page.
+RuleID: "Notion.PagePerms.APIPermsChanged"
+DisplayName: "Notion Page API Permissions Changed"
+Description: "A new API integration was added to a Notion page, or it's permissions were changed."
 LogTypes:
     - Notion.AuditLogs
 Enabled: true
