+++ rules/notion_rules/notion_page_accessible_to_guests.yml
@@ -1,0 +1,90 @@
+AnalysisType: rule
+RuleID: "Notion.PagePerms.GuestPermsChanged"
+DisplayName: "Notion Page Guest Permissions Changed"
+Description: The external guest permissions for a Notion page have been altered.
+LogTypes:
+    - Notion.AuditLogs
+Enabled: true
+Filename: notion_page_accessible_to_guests.py
+Runbook: Potential information exposure - review the shared page and rectify if needed.
+Severity: Low
+DedupPeriodMinutes: 60
+Threshold: 1
+Tags:
+  - Notion
+  - Public Exposure
+  - Beta
+Tests:
+  - Name: Guest Role Added
+    ExpectedResult: true
+    Log:
+      {
+        "event": {
+          "actor": {
+            "id": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx",
+            "object": "user",
+            "person": {
+              "email": "aragorn.elessar@lotr.com"
+            },
+            "type": "person"
+          },
+          "details": {
+            "entity": {
+              "id": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx",
+              "object": "user",
+              "person": {
+                "email": "frodo.baggins@lotr.com"
+              },
+              "type": "person"
+            },
+            "new_permission": "full_access",
+            "old_permission": "none",
+            "page_audience": "shared_internally",
+            "target": {
+              "page_id": "441356b5-557b-4053-8d2f-7932d2607d66",
+              "type": "page_id"
+            }
+          },
+          "id": "e18690f8-e24b-4b03-ba6f-123eb7ec0f08",
+          "timestamp": "2023-08-11 23:02:53.113000000",
+          "type": "page.permissions.guest_role_added",
+          "workspace_id": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
+        }
+      }
+
+  - Name: Guest Role Changed
+    ExpectedResult: true
+    Log:
+      {
+        "event": {
+          "actor": {
+            "id": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx",
+            "object": "user",
+            "person": {
+              "email": "aragorn.elessar@lotr.com"
+            },
+            "type": "person"
+          },
+          "details": {
+            "entity": {
+              "id": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx",
+              "object": "user",
+              "person": {
+                "email": "frodo.baggins@lotr.com"
+              },
+              "type": "person"
+            },
+            "new_permission": "full_access",
+            "old_permission": "read_only",
+            "page_audience": "shared_internally",
+            "target": {
+              "page_id": "441356b5-557b-4053-8d2f-7932d2607d66",
+              "type": "page_id"
+            }
+          },
+          "id": "e18690f8-e24b-4b03-ba6f-123eb7ec0f08",
+          "timestamp": "2023-08-11 23:02:53.113000000",
+          "type": "page.permissions.guest_role_updated",
+          "workspace_id": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
+        }
+      }
